Make parsing libraries private dependencies
