+++ test/test_partition_fwd_index.cpp
@@ -18,6 +18,7 @@
 #include "binary_freq_collection.hpp"
 #include "filesystem.hpp"
 #include "forward_index_builder.hpp"
+#include "parser.hpp"
 #include "payload_vector.hpp"
 #include "pisa_config.hpp"
 #include "sharding.hpp"
